FunctionalExpressionsIdentity inspection added, CollectionContainsUrl inspection modified
