+++ platform/platform-resources-en/src/inspectionDescriptions/FunctionalExpressionsIdentity.html
@@ -1,0 +1,9 @@
+<!-- Copyright 2000-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file. -->
+<html>
+<body>
+Reports objects which are a subtype of java.util.Set or java.util.Map that may contain keys of functional interface type.
+In such situations non-captured lambdas/method references could be stored
+in sets which could lead to unexpected results because lambdas can be reused.
+<p><small>New in 2019.1</small></p>
+</body>
+</html>
