+++ plugins/InspectionGadgets/InspectionGadgetsAnalysis/src/com/siyeh/InspectionGadgetsBundle.properties
@@ -1536,8 +1536,11 @@ enumeration.can.be.iteration.problem.descriptor=<code>#ref()</code> can be repla
 enumeration.can.be.iteration.quickfix=Replace with 'Iterator' construct
 equals.hashcode.called.on.url.display.name='equals()' or 'hashCode()' called on 'java.net.URL' object
 equals.hashcode.called.on.url.problem.descriptor=Call to <code>#ref()</code> on URL object #loc
-collection.contains.url.problem.decriptor={0} <code>#ref</code> may contain URL objects #loc
+collection.contains.url.problem.descriptor={0} <code>#ref</code> may contain URL objects #loc
 collection.contains.url.display.name=Map or Set may contain 'java.net.URL' objects
+ignore.func.interface.annotation.option=Ignore types with @FunctionalInterface
+functional.expression.identity=Collection has key of functional interface type
+collection.contains.key.problem.descriptor={0} <code>#ref</code> may contain key of functional interface type #loc
 implicit.array.to.string.problem.descriptor=Implicit call to 'toString()' on array <code>#ref</code> #loc
 explicit.array.to.string.problem.descriptor=Call to '#ref()' on array #loc
 implicit.array.to.string.method.call.problem.descriptor=Implicit call to 'toString()' on array returned by call to <code>#ref</code> #loc
